__len__() returns an int now

+++ starnet/data/utilities/generator.py
@@ -53,7 +53,7 @@ class DataGenerator(keras.utils.Sequence):
 
     def __len__(self):
         'Denotes the number of batches per epoch'
-        return np.floor(len(self.indices) // self.batch_size)
+        return int(np.floor(len(self.indices) // self.batch_size))
 
     def __getitem__(self, index):
         'Generate one batch of data'
